Abstract Debezium component functional tests
