+++ db_components/debezium/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Component tests for Debezium based CDC components
+"""
