+++ db_components/ex_postgres_cdc/tests/functional/simple-incremental-chained/02_inserts_updates_deletes/source/set_up.py
@@ -2,7 +2,7 @@ import os
 
 from datadirtest import TestDataDir
 
-from tests.test_functional import TestDatabaseEnvironment
+from db_components.debezium.tests.functional import TestDatabaseEnvironment
 
 
 def get_transactions_queries():
